+++ src/Console/Configuration.fs
@@ -9,6 +9,7 @@ module Configuration
     | [<CliPrefix(CliPrefix.None); First; Unique>] Up
     | [<CliPrefix(CliPrefix.None); First; Unique>] Down of count:uint
     | [<CliPrefix(CliPrefix.None); First; Unique>] List
+    | [<CliPrefix(CliPrefix.None); First; Unique>] New of file_name:string
     interface IArgParserTemplate with 
       member self.Usage =
         match self with
@@ -17,6 +18,7 @@ module Configuration
         | Up _ -> "(default) Run all outstanding UP migrations"
         | Down _ -> "Revert <count> of run migrations"
         | List _ -> "Displays a list of the un-migrated scripts still in the migration folder"
+        | New _ -> "Creates a new script file in the migrations folder"
 
 
   let configure argv =
@@ -33,7 +35,10 @@ module Configuration
       | None ->
         match (options.TryGetResult Down) with
         | Some d -> FSharp.Data.Migrations.Down d
-        | None -> FSharp.Data.Migrations.Action.Up
+        | None -> 
+          match (options.TryGetResult New) with
+          | Some f -> FSharp.Data.Migrations.New f
+          | None -> FSharp.Data.Migrations.Action.Up
 
     // Return a record with values or default's set
     {|